# docs/swagger.yaml
basePath: /api/v2/
definitions:
  handler.AddAuthor:
    properties:
      author_name:
        type: string
      group_id:
        type: string
    type: object
  handler.AddGroup:
    properties:
      group_name:
        type: string
    type: object
...
  handler.DeleteGroup:
    properties:
      group_id:
        type: string
    type: object
...
  handler.RemoveAuthor:
    properties:
      group_auther_id:
        type: string
    type: object
  handler.TelegramUpsert:
    properties:
      chat_id:
        type: string
      user_id:
        type: string
    type: object
...
  handler.UpdateGroupNameRequest:
    properties:
      group_id:
        type: string
      new_group_name:
        type: string
    type: object
...
info:
  contact:
    email: admintool@investicstudio.com
    name: API Support
  description: API documentation for Crypto Studio Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Studio Backend API
  version: "1.0"
paths:
...
  /notification/add-author:
    post:
      consumes:
      - application/json
      description: add new author to group
      parameters:
      - description: Add New Author
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddAuthor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: add new author to group godoc
      tags:
      - Notification
  /notification/add-group:
    post:
      consumes:
      - application/json
      description: Add new group
      parameters:
      - description: Add New Group
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add new group godoc
      tags:
      - Notification
  /notification/count-group:
    get:
      consumes:
      - application/json
      description: Count user's group
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Count user's group godoc
      tags:
      - Notification
  /notification/delete-group-author:
    post:
      consumes:
      - application/json
      description: Delete all author in group
      parameters:
      - description: Delete all author in group
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.DeleteGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete all author in group godoc
      tags:
      - Notification
  /notification/disconnect:
    get:
      consumes:
      - application/json
      description: Delete telegram data, group id and author
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete telegram data, group id and author godoc
      tags:
      - Notification
  /notification/get-group:
    get:
      consumes:
      - application/json
      description: Get Notification group list by uid
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Notification group list by uid
      tags:
      - Notification
  /notification/get-telegram:
    get:
      consumes:
      - application/json
      description: Get user telegram data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user telegram data godoc
      tags:
      - Notification
  /notification/remove-author:
    post:
      consumes:
      - application/json
      description: remove author from group
      parameters:
      - description: Remove Author
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RemoveAuthor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: remove author from group godoc
      tags:
      - Notification
  /notification/update-group-name:
    post:
      consumes:
      - application/json
      description: Update group name by group id
      parameters:
      - description: Update Group Name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateGroupNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update group name by group id
      tags:
      - Notification
  /notification/update-telegram:
    post:
      consumes:
      - application/json
      description: Update user telegram data if telegram user id and chat id are valid
      parameters:
      - description: Update Telegram Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.TelegramUpsert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user telegram data godoc
      tags:
      - Notification
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your-token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"